Class AdbiResult (and methods)

AdbiResult objects are created by DBI::dbSendQuery() or DBI::dbSendStatement(), and encapsulate the result of an SQL statement (either SELECT or not). They are a superclass of the DBI::DBIResult class. The "Usage" section lists the class methods overridden by adbi.

Usage

# S4 method for class 'AdbiResult'
dbBindArrow(res, params, ...)

# S4 method for class 'AdbiResult'
dbBind(res, params, ...)

# S4 method for class 'AdbiResult'
dbClearResult(res, ...)

# S4 method for class 'AdbiResult'
dbColumnInfo(res, ...)

# S4 method for class 'AdbiResult'
dbGetRowCount(res, ...)

# S4 method for class 'AdbiResult'
dbGetRowsAffected(res, ...)

# S4 method for class 'AdbiResult'
dbGetStatement(res, ...)

# S4 method for class 'AdbiResult'
dbHasCompleted(res, ...)

# S4 method for class 'AdbiResult'
dbIsValid(dbObj, ...)

# S4 method for class 'AdbiResult'
show(object)

Arguments

res

An object inheriting from DBI::DBIResult.

params

For dbBind(), a list of values, named or unnamed, or a data frame, with one element/column per query parameter. For dbBindArrow(), values as a nanoarrow stream, with one column per query parameter.

...

Other arguments passed on to methods.

dbObj

An object inheriting from DBI::DBIObject, i.e. DBI::DBIDriver, DBI::DBIConnection, or a DBI::DBIResult

object

Any R object

See also