Class AdbiResultArrow (and methods)

AdbiResultArrow objects are created by DBI::dbSendQueryArrow(), and encapsulate the result of an SQL query (a SELECT statement). They are a superclass of the DBI::DBIResultArrow class. The "Usage" section lists the class methods overridden by adbi.

Usage

# S4 method for class 'AdbiResultArrow'
dbBindArrow(res, params, ...)

# S4 method for class 'AdbiResultArrow'
dbBind(res, params, ...)

# S4 method for class 'AdbiResultArrow'
dbClearResult(res, ...)

# S4 method for class 'AdbiResultArrow'
dbColumnInfo(res, ...)

# S4 method for class 'AdbiResultArrow'
dbFetchArrowChunk(res, ...)

# S4 method for class 'AdbiResultArrow'
dbFetchArrow(res, ...)

# S4 method for class 'AdbiResultArrow'
dbGetRowCount(res, ...)

# S4 method for class 'AdbiResultArrow'
dbGetRowsAffected(res, ...)

# S4 method for class 'AdbiResultArrow'
dbGetStatement(res, ...)

# S4 method for class 'AdbiResultArrow'
dbHasCompleted(res, ...)

# S4 method for class 'AdbiResultArrow'
dbIsValid(dbObj, ...)

# S4 method for class 'AdbiResultArrow'
show(object)

Arguments

res

An object inheriting from DBI::DBIResult.

params

For dbBind(), a list of values, named or unnamed, or a data frame, with one element/column per query parameter. For dbBindArrow(), values as a nanoarrow stream, with one column per query parameter.

...

Other arguments passed on to methods.

dbObj

An object inheriting from DBI::DBIObject, i.e. DBI::DBIDriver, DBI::DBIConnection, or a DBI::DBIResult

object

Any R object

See also