DBI methods

Implementations of pure virtual functions defined in the DBI package.

# S4 method for AdbiConnection
dbAppendTable(conn, name, value, ..., row.names = NULL)

# S4 method for AdbiConnection
dbBegin(conn, ...)

# S4 method for AdbiResult
dbBind(res, params, ...)

# S4 method for AdbiResult
dbClearResult(res, ...)

# S4 method for KazamResultArrow
dbClearResult(res, ...)

# S4 method for AdbiResult
dbColumnInfo(res, ...)

# S4 method for AdbiConnection
dbCommit(conn, ...)

# S4 method for AdbiDriver
dbConnect(drv, ...)

# S4 method for AdbiConnection,ANY
dbDataType(dbObj, obj, ...)

# S4 method for AdbiDriver,ANY
dbDataType(dbObj, obj, ...)

# S4 method for AdbiDriver,list
dbDataType(dbObj, obj, ...)

# S4 method for AdbiConnection
dbDisconnect(conn, ...)

# S4 method for AdbiConnection,Id
dbExistsTable(conn, name, ...)

# S4 method for AdbiConnection,SQL
dbExistsTable(conn, name, ...)

# S4 method for AdbiConnection,character
dbExistsTable(conn, name, ...)

# S4 method for KazamResultArrow
dbFetchArrow(res, ...)

# S4 method for AdbiResult
dbFetch(res, n = -1, ...)

# S4 method for AdbiConnection
dbGetInfo(dbObj, ...)

# S4 method for AdbiDriver
dbGetInfo(dbObj, ...)

# S4 method for AdbiResult
dbGetInfo(dbObj, ...)

# S4 method for AdbiResult
dbGetRowCount(res, ...)

# S4 method for AdbiResult
dbGetRowsAffected(res, ...)

# S4 method for AdbiResult
dbGetStatement(res, ...)

# S4 method for AdbiResult
dbHasCompleted(res, ...)

# S4 method for AdbiConnection
dbIsValid(dbObj, ...)

# S4 method for AdbiDriver
dbIsValid(dbObj, ...)

# S4 method for AdbiResult
dbIsValid(dbObj, ...)

# S4 method for AdbiConnection,Id
dbListFields(conn, name, ...)

# S4 method for AdbiConnection,SQL
dbListFields(conn, name, ...)

# S4 method for AdbiConnection,character
dbListFields(conn, name, ...)

# S4 method for AdbiConnection
dbListTables(conn, ...)

# S4 method for AdbiConnection,character
dbQuoteIdentifier(conn, x, ...)

# S4 method for AdbiConnection,character
dbQuoteLiteral(conn, x, ...)

# S4 method for AdbiConnection,character
dbQuoteString(conn, x, ...)

# S4 method for AdbiConnection,character
dbRemoveTable(conn, name, ..., temporary = FALSE, fail_if_missing = TRUE)

# S4 method for AdbiConnection
dbRollback(conn, ...)

# S4 method for AdbiConnection
dbSendQueryArrow(conn, statement, ..., params = NULL)

# S4 method for AdbiConnection,character
dbSendQuery(conn, statement, ..., params = NULL, immediate = NULL)

# S4 method for AdbiConnection,character
dbSendStatement(conn, statement, ..., params = NULL, immediate = NULL)

# S4 method for AdbiConnection
dbUnquoteIdentifier(conn, x, ...)

# S4 method for AdbiConnection,Id,data.frame
dbWriteTable(
  conn,
  name,
  value,
  overwrite = FALSE,
  append = FALSE,
  ...,
  field.types = NULL,
  row.names = NULL,
  temporary = FALSE
)

# S4 method for AdbiConnection,SQL,data.frame
dbWriteTable(conn, name, value, ...)

# S4 method for AdbiConnection,character,data.frame
dbWriteTable(conn, name, value, ...)

# S4 method for AdbiConnection
show(object)

# S4 method for AdbiDriver
show(object)

# S4 method for AdbiResult
show(object)

Arguments

conn

A DBIConnection object, as returned by dbConnect().

name

The table name, passed on to dbQuoteIdentifier(). Options are:

a character string with the unquoted DBMS table name, e.g. "table_name",

a call to Id() with components to the fully qualified table name, e.g. Id(schema = "my_schema", table = "table_name")

a call to SQL() with the quoted and fully qualified table name given verbatim, e.g. SQL('"my_schema"."table_name"')

value

For dbWriteTable(), a data.frame (or coercible to data.frame). For dbWriteTableArrow(), an object coercible to an Arrow RecordBatchReader.

...

Other parameters passed on to methods.

row.names

A logical specifying whether the row.names should be output to the output DBMS table; if TRUE, an extra field whose name will be whatever the R identifier "row.names" maps to the DBMS (see DBI::make.db.names()). If NA will add rows names if they are characters, otherwise will ignore.

res

An object inheriting from DBIResult.

params

A list of bindings, named or unnamed.

drv

an object that inherits from DBIDriver, or an existing DBIConnection object (in order to clone an existing connection).

dbObj

A object inheriting from DBIDriver or DBIConnection

obj

An R object whose SQL type we want to determine.

n

maximum number of records to retrieve per fetch. Use n = -1 or n = Inf to retrieve all pending records. Some implementations may recognize other special values.

x

A character vector, SQL or Id object to quote as identifier.

temporary

a logical specifying whether the new table should be temporary. Its default is FALSE.

fail_if_missing

If FALSE, dbRemoveTable() succeeds if the table doesn't exist.

statement

a character string containing SQL.

immediate

Passing a value TRUE is intended for statements containing no placeholders and FALSE otherwise. The default value NULL will inspect the statement for presence of placeholders (will PREPARE the statement)

overwrite

Allow overwriting the destination table. Cannot be TRUE if append is also TRUE.

append

Allow appending to the destination table. Cannot be TRUE if overwrite is also TRUE.

field.types

character vector of named SQL field types where the names are the names of new table's columns. If missing, types inferred with DBI::dbDataType()).

object

Any R object