Class AdbiResult (and methods)

AdbiResult objects are created by DBI::dbSendQuery() or DBI::dbSendStatement(), and encapsulate the result of an SQL statement (either SELECT or not). They are a superclass of the DBIResult class. The "Usage" section lists the class methods overridden by adbi.

# S4 method for AdbiResult
dbBindArrow(res, params, ...)

# S4 method for AdbiResult
dbBind(res, params, ...)

# S4 method for AdbiResult
dbClearResult(res, ...)

# S4 method for AdbiResult
dbColumnInfo(res, ...)

# S4 method for AdbiResult
dbGetRowCount(res, ...)

# S4 method for AdbiResult
dbGetRowsAffected(res, ...)

# S4 method for AdbiResult
dbGetStatement(res, ...)

# S4 method for AdbiResult
dbHasCompleted(res, ...)

# S4 method for AdbiResult
dbIsValid(dbObj, ...)

# S4 method for AdbiResult
show(object)

Arguments

res

An object inheriting from DBIResult.

params

For dbBind(), a list of values, named or unnamed, or a data frame, with one element/column per query parameter. For dbBindArrow(), values as a nanoarrow stream, with one column per query parameter.

...

Other arguments passed on to methods.

dbObj

An object inheriting from DBIObject, i.e. DBIDriver, DBIConnection, or a DBIResult

object

Any R object

See also