Class AdbiResultArrow (and methods)

AdbiResultArrow objects are created by DBI::dbSendQueryArrow(), and encapsulate the result of an SQL query (a SELECT statement). They are a superclass of the DBIResultArrow class. The "Usage" section lists the class methods overridden by adbi.

Usage

# S4 method for AdbiResultArrow
dbBindArrow(res, params, ...)

# S4 method for AdbiResultArrow
dbBind(res, params, ...)

# S4 method for AdbiResultArrow
dbClearResult(res, ...)

# S4 method for AdbiResultArrow
dbColumnInfo(res, ...)

# S4 method for AdbiResultArrow
dbFetchArrowChunk(res, ...)

# S4 method for AdbiResultArrow
dbFetchArrow(res, ...)

# S4 method for AdbiResultArrow
dbGetRowCount(res, ...)

# S4 method for AdbiResultArrow
dbGetRowsAffected(res, ...)

# S4 method for AdbiResultArrow
dbGetStatement(res, ...)

# S4 method for AdbiResultArrow
dbHasCompleted(res, ...)

# S4 method for AdbiResultArrow
dbIsValid(dbObj, ...)

# S4 method for AdbiResultArrow
show(object)

Arguments

res

An object inheriting from DBIResult.

params

For dbBind(), a list of values, named or unnamed, or a data frame, with one element/column per query parameter. For dbBindArrow(), values as a nanoarrow stream, with one column per query parameter.

...

Other arguments passed on to methods.

dbObj

An object inheriting from DBIObject, i.e. DBIDriver, DBIConnection, or a DBIResult

object

Any R object

See also